Split icu_provider_baked into a regular and exporter crate
