+++ provider/baked_export/README.md
@@ -1,0 +1,11 @@
+# icu_provider_baked_export [![crates.io](https://img.shields.io/crates/v/icu_provider_baked_export)](https://crates.io/crates/icu_provider_baked_export)
+
+<!-- cargo-rdme start -->
+
+Tooling for the baked provider.
+
+<!-- cargo-rdme end -->
+
+## More Information
+
+For more information on development, authorship, contributing etc. please visit [`ICU4X home page`](https://github.com/unicode-org/icu4x).
